+++ tests/poll/test_models.py
@@ -4,7 +4,8 @@ import re
 import requests
 import requests_mock
 
-from bulbs.poll.models import Poll, Answer, SodaheadResponseError
+from bulbs.poll.models import Poll, Answer
+from bulbs.poll.sodahead import SodaheadResponseError
 from bulbs.utils.test import (
     BaseIndexableTestCase,
     random_title,
